Portal/requirements.txt
asgiref
certifi
Django==4.0.1
django-auth-ldap
django-cors-headers
django-middleware-global-request
djangorestframework
idna
jalali_core
jdatetime
mssql-django
pillow
pyasn1
pyasn1_modules
PyJWT
pyodbc
python-dotenv
.\python_ldap-3.4.0-cp39-cp39-win_amd64.whl
pytz
requests
sqlparse
typing_extensions
tzdata
urllib3
